C2LC-664: Fix auto-scrolling to moved loop blocks
Add the program step number to the React 'key' for loop sections
of the rendered program.

This change causes React to rebuild loop sections when a loop is
moved and restores the automatic scrolling to a loop block when it is
moved off-screen using the move next and move previous operations.

Automatic scrolling to moved loop blocks stopped working due to the
changes in commit caa3989ec588dfb2ee64e8efa55a195945123bc8