C2LC-668: Cache voices and listen for changes
Retrieve the list of speech synthesis voices at construction of
AudioManagerImpl and register a handler to update the list when the
voices change.

This is to fix an issue with Firefox where an empty list could be
returned from the first call to getVoices() but a populated list is
returned from subsequent calls. This change in the list of voices
could cause the voice that the user hears to change if the voice
selection logic implemented for C2LC-668 differs from the browser's
built-in voice selection logic.