C2LC-683: Set background-color by world
This is to reduce the flicker/flash when Weavly is reloaded. Without
this change, the scene background would be white until the background
image for the current world is loaded and then it would change to the
background image. When Weavly is reloaded this causes the scene
background to flash white. With this change, each world and theme
combination has a background colour, which is visible until the world
background is available. Instead of white, the scene background colour
will match the background image more closely.