C2LC-681: Add 'alwaysShowVerticalScrollbar' prop to ModalBody
This prop switches the ModalBody from 'overflow-y: auto' to
'overflow-y: scroll'.

This new prop is used by the WorldSelector to ensure we get a consistent
width for the dialog across browsers. Including letting us set
'width: fit-content' on the options grid to ensure that the grey
background is displayed across the whole dialog when the width is
reduced to cause a horizontal scrollbar to be necessary.